+++ spatialMeanFilter.m
@@ -1,4 +1,4 @@
 function result = spatialMeanFilter(f, n)
     mask = ones(n, n) / (n * n);
-    result = konvolusiCitra(f, mask, floor(n/2), 1);
+    result = uint8(konvolusiCitra(double(f), mask, floor(n/2), 1));
 end
